Moved permission helper functions out of the views
